Extract couchdb package
